+++ GatePe/Facilities/views.py
@@ -10,50 +10,69 @@ class FacilityView(APIView):
     serializer_class = FacilitySerializer
     
     def get(self, request):
-        response = {'status': 200}
-        display_facility = Facility.objects.all()
-        serializer = FacilitySerializer(display_facility, many= True)
-        response['data'] = serializer.data
-        return Response(response)
-    
-    def post(self, request):
-        data = request.data
-        serializer = FacilitySerializer(data = data)
-        if serializer.is_valid():
-            serializer.save()
-            response = {"status":200, "message":"Data added"}
+        if request.user_id:
+
+            response = {'status': 200}
+            display_facility = Facility.objects.all()
+            serializer = FacilitySerializer(display_facility, many= True)
+            response['data'] = serializer.data
             return Response(response)
-        return Response({"Status":400, "message":"Data not found"})
+        else:
+            return Response({'Error':'Unauthorized'}) 
+
+
+    def post(self, request):
+        if request.user_id:
+
         
-    def delete(self, request):
-        data = request.data
-        try:
-            obj = Facility.objects.get(id=data.get('id'))
-            obj.delete()
-            return Response({"Status":200, "message":"Deleted"})
-        except Exception as e:
-            print(e)
-        return Response({"Status":400, "message":"Data not found"})
-    
-     
-    def patch(self, request):
-        
-        response = {"status":200}
-        data = request.data
-        try:
-            obj = Facility.objects.get(id=data.get('id'))
-            serializer = FacilitySerializer(obj, data=data, partial = True)
+            data = request.data
+            serializer = FacilitySerializer(data = data)
             if serializer.is_valid():
                 serializer.save()
-                response['data'] = serializer.data
+                response = {"status":200, "message":"Data added"}
                 return Response(response)
-            else:
-                return Response(serializer.error)
-        except Exception as e:
-            print(e)
-        return Response({'Status':400, "message":'Invalid id'})
-    
-    
+            return Response({"Status":400, "message":"Data not found"})
+        else:
+            return Response({'Error':'Unauthorized'}) 
+ 
+    def delete(self, request):
+        if request.user_id:
+
+            data = request.data
+            try:
+                obj = Facility.objects.get(id=data.get('id'))
+                obj.delete()
+                return Response({"Status":200, "message":"Deleted"})
+            except Exception as e:
+                print(e)
+            return Response({"Status":400, "message":"Data not found"})
+        else:
+            return Response({'Error':'Unauthorized'}) 
+
+
+     
+    def patch(self, request):
+        if request.user_id:
+
+        
+            response = {"status":200}
+            data = request.data
+            try:
+                obj = Facility.objects.get(id=data.get('id'))
+                serializer = FacilitySerializer(obj, data=data, partial = True)
+                if serializer.is_valid():
+                    serializer.save()
+                    response['data'] = serializer.data
+                    return Response(response)
+                else:
+                    return Response(serializer.error)
+            except Exception as e:
+                print(e)
+            return Response({'Status':400, "message":'Invalid id'})
+        else:
+            
+            return Response({'Error':'Unauthorized'}) 
+
 
 # Billing view
 class BillingView(APIView):
@@ -61,51 +80,68 @@ class BillingView(APIView):
     serializer_class = BillingSerializer
     
     def get(self, request):
-        response = {'status': 200}
-        display_billing = Billing.objects.all()
-        serializer = BillingSerializer(display_billing, many= True)
-        response['data'] = serializer.data
-        return Response(response)
-    
-    def post(self, request):
-        data = request.data
-        serializer = BillingSerializer(data = data)
-        if serializer.is_valid():
-            serializer.save()
-            response = {"status":200, "message":"Data added"}
+        if request.user_id:
+
+            response = {'status': 200}
+            display_billing = Billing.objects.all()
+            serializer = BillingSerializer(display_billing, many= True)
+            response['data'] = serializer.data
             return Response(response)
-        return Response({"Status":400, "message":"Data not found"})
-        
-    def delete(self, request, ):
-        
-        response = {"status":200}
-        data = request.data
-        try:
-            obj = Billing.objects.get(id=data.get('id'))
-            obj.delete()
-            return Response({"Status":200, "mesage":"Deleted"})
-        except Exception as e:
-            print(e)
-        return Response({"Status":400, "message":"Data not found"})
-    
-     
-    def patch(self, request):
-        
-        response = {"status":200}
-        data = request.data
-        try:
-            obj = Billing.objects.get(id=data.get('id'))
-            serializer = BillingSerializer(obj, data=data, partial = True)
+        else:
+            return Response({'Error':'Unauthorized'}) 
+
+
+    def post(self, request):
+        if request.user_id:
+
+            data = request.data
+            serializer = BillingSerializer(data = data)
             if serializer.is_valid():
                 serializer.save()
-                response['data'] = serializer.data
+                response = {"status":200, "message":"Data added"}
                 return Response(response)
-            else:
-                return Response(serializer.error)
-        except Exception as e:
-            print(e)
-        return Response({'Status':400, "message":'Invalid id'})
-    
+            return Response({"Status":400, "message":"Data not found"})
+        else:
+            return Response({'Error':'Unauthorized'}) 
+  
+    def delete(self, request, ):
+        if request.user_id:
+
+        
+            response = {"status":200}
+            data = request.data
+            try:
+                obj = Billing.objects.get(id=data.get('id'))
+                obj.delete()
+                return Response({"Status":200, "mesage":"Deleted"})
+            except Exception as e:
+                print(e)
+            return Response({"Status":400, "message":"Data not found"})
+        else:
+            return Response({'Error':'Unauthorized'}) 
+
+
+     
+    def patch(self, request):
+        if request.user_id:
+
+            response = {"status":200}
+            data = request.data
+            try:
+                obj = Billing.objects.get(id=data.get('id'))
+                serializer = BillingSerializer(obj, data=data, partial = True)
+                if serializer.is_valid():
+                    serializer.save()
+                    response['data'] = serializer.data
+                    return Response(response)
+                else:
+                    return Response(serializer.error)
+            except Exception as e:
+                print(e)
+            return Response({'Status':400, "message":'Invalid id'})
+        else:
+            return Response({'Error':'Unauthorized'}) 
+
 
 # Membership view
 class MembershipView(APIView):
@@ -113,47 +149,63 @@ class MembershipView(APIView):
     serializer_class = MembershipSerializer
     
     def get(self, request):
-        response = {'status': 200}
-        display_membership = Membership.objects.all()
-        serializer = MembershipSerializer(display_membership, many= True)
-        response['data'] = serializer.data
-        return Response(response)
-    
-    def post(self, request):
-        data = request.data
-        serializer = MembershipSerializer(data = data)
-        if serializer.is_valid():
-            serializer.save()
-            response = {"status":200, "message":"Data added"}
+        if request.user_id:
+
+            response = {'status': 200}
+            display_membership = Membership.objects.all()
+            serializer = MembershipSerializer(display_membership, many= True)
+            response['data'] = serializer.data
             return Response(response)
-        return Response({"Status":400, "message":"Data not found"})
-        
-    def delete(self, request, ):
-        
-        response = {"status":200}
-        data = request.data
-        try:
-            obj = Membership.objects.get(id=data.get('id'))
-            obj.delete()
-            return Response({"Status":200, "mesage":"Deleted"})
-        except Exception as e:
-            print(e)
-        return Response({"Status":400, "message":"Data not found"})
-    
-     
-    def patch(self, request, *args, **kwargs):
-        
-        response = {"status":200}
-        data = request.data
-        try:
-            obj = Membership.objects.get(id=data.get('id'))
-            serializer = MembershipSerializer(obj, data=data, partial = True)
+        else:
+            return Response({'Error':'Unauthorized'}) 
+
+    def post(self, request):
+        if request.user_id:
+
+            data = request.data
+            serializer = MembershipSerializer(data = data)
             if serializer.is_valid():
                 serializer.save()
-                response['data'] = serializer.data
+                response = {"status":200, "message":"Data added"}
                 return Response(response)
-            else:
-                return Response(serializer.error)
-        except Exception as e:
-            print(e)
-        return Response({'Status':400, "message":'Invalid id'})
+            return Response({"Status":400, "message":"Data not found"})
+        else:
+            return Response({'Error':'Unauthorized'}) 
+
+   
+    def delete(self, request, ):
+        if request.user_id:
+
+            response = {"status":200}
+            data = request.data
+            try:
+                obj = Membership.objects.get(id=data.get('id'))
+                obj.delete()
+                return Response({"Status":200, "mesage":"Deleted"})
+            except Exception as e:
+                print(e)
+            return Response({"Status":400, "message":"Data not found"})
+        else:
+            return Response({'Error':'Unauthorized'}) 
+
+     
+    def patch(self, request, *args, **kwargs):
+        if request.user_id:
+
+        
+            response = {"status":200}
+            data = request.data
+            try:
+                obj = Membership.objects.get(id=data.get('id'))
+                serializer = MembershipSerializer(obj, data=data, partial = True)
+                if serializer.is_valid():
+                    serializer.save()
+                    response['data'] = serializer.data
+                    return Response(response)
+                else:
+                    return Response(serializer.error)
+            except Exception as e:
+                print(e)
+            return Response({'Status':400, "message":'Invalid id'})
+        else:
+            return Response({'Error':'Unauthorized'})
